backend/requirements.txt
django>=5.0.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
